requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
tldextract==5.1.2
publicsuffix2==2.20191221
manuf==1.1.5
ipwhois==1.2.0
dnspython==2.0.0
scapy==2.5.0
pyshark==0.6.0
